extract breakpoint data into separate module
